feat: 현재 선택된 오늘의 할일을 나타내는 상태 추가
