+++ src/components/molecules/TodayTodoBox/index.tsx
@@ -1,5 +1,3 @@
-import { useState } from 'react';
-
 import { convertTime } from '@/utils/time';
 
 import { TodoDataTypes } from '@/types/userData';
